--- common/components/Header/data.tsx
@@ -1,16 +1,0 @@
-import { PageCategoryType } from "@/constants/globalTypes";
-
-// 隱藏「我要開團」的頁面
-export const hiddenGroupBtnPages: Array<string> = [
-  "login",
-  "signup",
-  "forget-password",
-  "create-group",
-];
-
-// 隱藏「在揪遊上成立店家」的頁面
-export const hiddenStoreBtnPages: Array<string> = ["login"];
-
-export type HeaderProps = {
-  pageCategory: PageCategoryType;
-};
